requirements.txt
netaddr==0.7.18
PyYAML==3.11
requests==2.9.1
